split constants
