+++ src/constants/electron.js
@@ -1,0 +1,7 @@
+const ELECTRON_SECOND_INSTANCE = 'second-instance'
+const ELECTRON_WINDOW_ALL_CLOSED = 'window-all-closed'
+
+module.exports = {
+  ELECTRON_SECOND_INSTANCE,
+  ELECTRON_WINDOW_ALL_CLOSED
+}
