+++ src/constants/ipc.js
@@ -1,0 +1,25 @@
+const IPC_RESIZE = 'resize'
+const IPC_DIALOG = 'dialog'
+const IPC_LAUNCH = 'launch'
+const IPC_THROW = 'throw'
+const IPC_THROW_FAR = 'throwFar'
+const IPC_MOVE_MONITOR = 'moveMonitor'
+const IPC_STOP = 'stop'
+const IPC_ALIGN = 'align'
+const IPC_MOVE = 'move'
+const IPC_REQUEST_FOCUS = 'requestFocus'
+const IPC_SHOW_MENU = 'showMenu'
+
+module.exports = {
+  IPC_RESIZE,
+  IPC_DIALOG,
+  IPC_LAUNCH,
+  IPC_THROW,
+  IPC_THROW_FAR,
+  IPC_MOVE_MONITOR,
+  IPC_STOP,
+  IPC_ALIGN,
+  IPC_MOVE,
+  IPC_REQUEST_FOCUS,
+  IPC_SHOW_MENU,
+}
