+++ src/constants/plain.js
@@ -1,0 +1,19 @@
+const OWL_FILES = ['climb-0.png', 'climb-1.png', 'fall-0.png', 'fall-1.png', 'sit-0.png', 'stand-0.png', 'tilt-0.png', 'tilt-1.png', 'walk-0.png', 'walk-1.png', 'walk-2.png', 'walk-3.png']
+const BASE64_META = 'data:image/png;base64, '
+const BEHAVIOR_JSON = 'behaviors.json'
+const CUSTOM_FILE_EXTENSION = '.ptc'
+const CURRENT_DIRECTORY = '.'
+const NUMBER = 'number'
+const THROW_DENOM = 30
+const ALWAYS_ON_TOP_LEVEL = 'pop-up-menu'
+
+module.exports = {
+  OWL_FILES,
+  BASE64_META,
+  BEHAVIOR_JSON,
+  NUMBER,
+  THROW_DENOM,
+  CUSTOM_FILE_EXTENSION,
+  CURRENT_DIRECTORY,
+  ALWAYS_ON_TOP_LEVEL,
+}
