+++ src/constants/text.js
@@ -1,0 +1,5 @@
+const FILE_NOT_VALID = '유효하지 않은 파일이에요.'
+
+module.exports = {
+  FILE_NOT_VALID,
+}
